[refactor] 유저쿠폰 도메인 생성에 따른 유저 도메인 리팩토링 및 생성
